refactor(webpack): extract processing modules
