+++ packages/webpack/src/plugin.js
@@ -1,3 +1,5 @@
+const { createModulesAnalyzer } = require('./buildtime/modulesData.js')
+
 const path = require('node:path')
 const assert = require('node:assert')
 
@@ -209,51 +211,17 @@ class LavaMoatPlugin {
         // processing of the paths involved in the bundle and cross-check with policy
 
         /**
-         * Array of strings representing the excludes found in the generation
-         * process.
-         *
-         * @type {string[]}
-         */
-        const excludes = []
-        /**
-         * Array of objects representing the paths and module ids found in the
-         * generation process.
-         *
-         * @type {{ path: string; moduleId: string | number }[]}
-         */
-        const knownPaths = []
-        /**
-         * Array of module ids that are unenforceable by policy.
-         *
-         * @type {(string | number)[]}
-         */
-        const unenforceableModuleIds = []
-        /**
-         * Array of module ids that are context modules and need to be
-         * double-wrapped.
-         *
-         * @type {{ moduleId: string | number; context: string }[]}
-         */
-        const contextModules = []
-
-        /**
          * Array of chunk ids that have been processed.
          *
          * @type {(string | number)[]}
          */
         const chunkIds = []
-        /**
-         * A record of module ids that are externals and need to be enforced as
-         * builtins.
-         *
-         * @type {Record<string | number, string>}
-         */
-        const externals = {}
+
         /**
          * @type {import('./buildtime/aa.js').IdentifierLookup}
          */
         let identifierLookup
-        const runChecks = this.options.runChecks
+        const runChecks = options.runChecks
 
         // Caveat: this might be called before the lookup map is ready if a plugin is running a child compilation or alike.
         // Note that in those cases wrapped code is not meant to run and policy will be empty.
@@ -265,94 +233,10 @@ class LavaMoatPlugin {
           return identifierLookup.pathToResourceId(p)
         }
 
-        const coveredTypes = [
-          JAVASCRIPT_MODULE_TYPE_AUTO,
-          JAVASCRIPT_MODULE_TYPE_DYNAMIC,
-          JAVASCRIPT_MODULE_TYPE_ESM,
-        ]
+        const moduleAnalyzer = createModulesAnalyzer({
+          mainCompilationWarnings,
+        })
 
-        /**
-         * @remarks
-         * Webpack has a concept of ignored modules. When a module is ignored, a
-         * carve-out is necessary in policy enforcement for it because the ID
-         * that webpack creates for it is not exactly helpful. example outcome
-         * in the bundle: `const nodeCrypto = __webpack_require__(/*! crypto *\/
-         * "?0b7d");` Sadly, even treeshaking doesn't eliminate that module.
-         * It's left there and failing to work when reached by runtime policy
-         * enforcement. Below is the most reliable way I've found to date to
-         * identify ignored modules.
-         * @param {import('webpack').Module} m
-         * @returns {boolean}
-         */
-        const isIgnoredModule = (m) => {
-          return Boolean(
-            m.type === JAVASCRIPT_MODULE_TYPE_DYNAMIC &&
-              // @ts-expect-error BAD TYPES
-              m.identifierStr?.startsWith('ignored')
-          )
-        }
-
-        /**
-         * Checks if a module is a context module.
-         *
-         * @param {any} m - The module to check.
-         * @param {string} moduleClass - The class of the module.
-         * @returns {boolean} - Returns true if the module is a context module,
-         *   otherwise false.
-         */
-        const isContextModule = (m, moduleClass) =>
-          moduleClass === 'ContextModule'
-
-        /**
-         * Identifies an asset that webpack includes in dist by default without
-         * setting any loaders explicitly.
-         *
-         * @param {import('webpack').Module} m
-         * @returns {m is import('webpack').NormalModule}
-         */
-        const isAmbientAsset = (m) =>
-          m.type === 'asset/resource' &&
-          'resource' in m &&
-          'loaders' in m &&
-          Array.isArray(m.loaders) &&
-          m.loaders.length === 0
-
-        /**
-         * @param {import('webpack').Module} m
-         * @param {string} moduleClass
-         * @returns {m is import('webpack').ExternalModule} // TODO: this is not
-         *   true anymore, but there's no superclass of all reasonable module
-         *   types
-         */
-        const isExternalModule = (m, moduleClass) =>
-          ['ExternalModule'].includes(moduleClass) &&
-          'externalType' in m &&
-          m.externalType !== undefined
-        /**
-         * @param {import('webpack').Module} m
-         * @param {string} moduleClass
-         * @returns {m is import('./buildtime/policyGenerator.js').InspectableWebpackModule}
-         */
-        const isInspectableModule = (m, moduleClass) =>
-          'userRequest' in m ||
-          m.type?.startsWith('javascript') ||
-          isExternalModule(m, moduleClass)
-
-        // Old: good for collecting all possible paths, but bad for matching them with module ids
-        // collect all paths resolved for the bundle and transition afterwards
-        // normalModuleFactory.hooks.afterResolve.tap(
-        //   PLUGIN_NAME,
-        //   (resolveData) => {
-        //     // TODO - typescript claims createData could be undefined. Do we care for those cases?
-        //     // Leaving it in a state where we'll get an error first time it happens.
-        //     if (coveredTypes.includes(resolveData.createData.type)) {
-        //       knownPaths.push([resolveData.createData.resourceResolveData.path, /*resolveData.createData.moduleId*/]);
-        //     }
-        //   }
-        // );
-        // compilation.hooks.finishModules.tap(PLUGIN_NAME, () => {
-        //   PROGRESS.report("pathsCollected");
-        // });
         compilation.hooks.afterOptimizeChunkIds.tap(PLUGIN_NAME, (chunks) => {
           try {
             const chunkGraph = compilation.chunkGraph
@@ -365,97 +249,22 @@ class LavaMoatPlugin {
 
               chunkGraph.getChunkModules(chunk).forEach((module) => {
                 const moduleId = chunkGraph.getModuleId(module)
-                const moduleClass =
-                  Object.getPrototypeOf(module).constructor.name
-
-                if (moduleId === null) {
-                  diag.rawDebug(
-                    2,
-                    `LavaMoatPlugin: module ${module.identifier()} has no moduleId, cannot cover it with policy.`
-                  )
-                  diag.rawDebug(4, { module })
-                  return
-                }
-
-                // Fixes the issue with assets being emitted to dist without a loader
-                // TODO: refactor to move random hardening of the build somewhere it's easier to track.
-                if (
-                  isAmbientAsset(module) &&
-                  module.resource.includes('node_modules') // FIXME: would be better to use canonicalName lookup and match with root
-                ) {
-                  // add a warning about removing the asset
-                  mainCompilationWarnings.push(
-                    new WebpackError(
-                      `LavaMoatPlugin: the following resource was being silently emitted to the dist directory and LavaMoat has prevented it: '${module.resource}'. If you want to add this resource, explicitly define a file-loader for it in your webpack configuration.`
-                    )
-                  )
-
-                  // We can't use `chunkGraph.disconnectChunkAndModule` here
-                  // because the require statement remains and errors out
-
-                  if (module.generatorOptions) {
-                    // generatorOptions was not present in testing, but types indicate it might be there
-                    module.generatorOptions.emit = false
-                  }
-                  if (module.generator) {
-                    module.generator = Object.create(module.generator, {
-                      emit: {
-                        value: false,
-                        writable: false,
-                        configurable: false,
-                        enumerable: true,
-                      },
-                    })
-                  }
-                }
-
-                // Note: module.context on an empty context module when no context information was guessable from code is going to point to the module that loads it.
-                if (isContextModule(module, moduleClass)) {
-                  diag.rawDebug(3, {
-                    contextModule: {
-                      moduleId,
-                      context: module.context,
-                      // @ts-expect-error we want to see it if available
-                      request: module?.options?.request,
-                      // @ts-expect-error we want to see it if available
-                      _identifier: module?._identifier,
-                    },
-                  })
-                  if (!module.context) {
-                    mainCompilationWarnings.push(
-                      new WebpackError(
-                        `LavaMoatPlugin: context module ${moduleId} has no context information. It cannot be allowed to work if it's reached at runtime.`
-                      )
-                    )
-                  } else {
-                    contextModules.push({ moduleId, context: module.context })
-                  }
-                }
-                if (isIgnoredModule(module)) {
-                  unenforceableModuleIds.push(moduleId)
-                } else {
-                  if (isExternalModule(module, moduleClass)) {
-                    externals[moduleId] = module.userRequest
-                  }
-                  if (isInspectableModule(module, moduleClass)) {
-                    policyGenerator.inspectWebpackModule(
-                      module,
-                      compilation.moduleGraph.getOutgoingConnections(module)
-                    )
-                  }
-
-                  // typescript is complaining about the use of `resource` here, but it's actually there.
-                  knownPaths.push({
-                    path: /** @type {any} */ (module).resource,
-                    moduleId,
-                  })
-                }
+                moduleAnalyzer.processModule({ module, moduleId })
               })
             })
-            diag.rawDebug(4, { knownPaths })
+
+            const moduleData = moduleAnalyzer.getResults()
+            diag.rawDebug(4, { knownPaths: moduleData.knownPaths })
             PROGRESS.report('pathsCollected')
 
             diag.rawDebug(2, 'writing policy')
+
+            for (const moduleToInspect of moduleData.inspectable) {
+              policyGenerator.inspectWebpackModule(
+                moduleToInspect,
+                compilation.moduleGraph.getOutgoingConnections(moduleToInspect)
+              )
+            }
             // use the generated policy to save the user one additional pass
             // getting the policy also writes all files where necessary
             const policyToApply = policyGenerator.getPolicy()
@@ -467,20 +276,20 @@ class LavaMoatPlugin {
 
             identifierLookup = generateIdentifierLookup({
               readableResourceIds: options.readableResourceIds,
-              unenforceableModuleIds,
-              contextModules,
-              externals,
-              paths: knownPaths,
+              unenforceableModuleIds: moduleData.unenforceableModuleIds,
+              contextModules: moduleData.contextModules,
+              externals: moduleData.externals,
+              paths: moduleData.knownPaths,
               policy: policyToApply,
               canonicalNameMap,
             })
 
             // =================================================================
 
-            if (unenforceableModuleIds.length > 0) {
+            if (moduleData.unenforceableModuleIds.length > 0) {
               mainCompilationWarnings.push(
                 new WebpackError(
-                  `LavaMoatPlugin: the following module ids can't be controlled by policy and must be ignored at runtime: \n  ${unenforceableModuleIds.join()}`
+                  `LavaMoatPlugin: the following module ids can't be controlled by policy and must be ignored at runtime: \n  ${moduleData.unenforceableModuleIds.join()}`
                 )
               )
             }
@@ -493,6 +302,19 @@ class LavaMoatPlugin {
 
         // =================================================================
         // javascript modules generator tweaks installation
+        /**
+         * Array of strings representing the excludes found in the generation
+         * process.
+         *
+         * @type {string[]}
+         */
+        const excludes = []
+
+        const coveredTypes = [
+          JAVASCRIPT_MODULE_TYPE_AUTO,
+          JAVASCRIPT_MODULE_TYPE_DYNAMIC,
+          JAVASCRIPT_MODULE_TYPE_ESM,
+        ]
 
         for (const moduleType of coveredTypes) {
           normalModuleFactory.hooks.generator.for(moduleType).tap(
